Update: libcommuni to 3.6.0
Pull in the updated files from upstream. Note that upstream does not
provide a CMake project file so it is not currently possible to import the
whole libcommuni as a git submodule. As such this does not currently work
for the CMake case as it is necessary to have either `IRC_STATIC` or
`IRC_SHARED` defined but I have not established how to do that for that
library.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>