Fix bug when unhighlighting during control-selection. (#4322)
Calling normaliseSelection during unHighlight() is incorrect, because
mPA and mPB are not necessarily equal to mDragStart and
mDragSelectionEnd in any order; during control-selection, mPA and mPB
are instead the beginning/end of the corresponding line.  So without
this, unhighlighting after control-selection leaves some bits
highlighted.

And we don't need to call normaliseSelection, because mPA and mPB are
always normalised when set.