Update: libcommuni to 3.6.0 (#4303)
Pull in the updated files from upstream. Note that upstream does not
provide a CMake project file so it is not currently possible to import the
whole libcommuni as a git submodule. As such this does not currently work
for the CMake case as it is necessary to have either `IRC_STATIC` or
`IRC_SHARED` defined but I have not established how to do that for that
library.

A tweak was needed for some CMakeLists.txt entries for libcommuni:
effectively adds a missing `include/IrcCore/irccore_p.h` from the list of
headers - so it shows up in Qt Creator listing.

Adjusted a define for the compilation of libcommuni library: having
tried `PRIVATE` and `INTERFACE` settings determine that it is
necessary to use `PUBLIC` for the definition of `IRC_STATIC` in the CMake
project file we (created and) use for libcommuni.

Update libcommuni date of Copyright details

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>