Cleanup: replace deprecated std::uncaught_exception()
`std::uncaught_exception()` was deprecated in C++17 and is to be removed in
C++20.

It has been replaced by `std::uncaught_exception()` which returns an `int`
that indicates how many exceptions are outstanding rather than the previous
`true` or `false` if there is any. This is helpful in situations where
exceptions could happen in exception handling code. That is a whole level
of brain-ache I do not even want to think about. However it provides a
perfectly suitable solution to our usage if we just drop it in as a
replacement...

For me this reduces the warning count from 662 to 510 (i.e. down by 152).

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>