Support for selecting words/lines with mouse by double/triple click (#4323)
* Support for selecting words/lines with mouse by double/triple click

Previously, you could double-click on a word to select it, but you
couldn't drag to increase the selection.  With this patch, if you
double click and drag, it selects a region delimited by word
boundaries.  If you triple-click and drag, it's delimited by lines
instead (same as ctrl-click).

Also fixes a bug in how the word selection was implemented:
previously, spaces and user-defined delimiters were handled
separately, and would incorrectly not delimit words if no spaces were
adjacent (e.g., it would select the entirety of foo""bar ).

* Edits from review

* Explanation of mMouseTrackLevel