Revise: restore profile name to connection icon
This is so that the keyboard selection of profiles beginning with a
particular letter works again. This will close issue #4283 .

This also returns it to a readable size instead of trying to hide it by
setting the font size to 1 - as that would still show something (and a font
size of zero is NOT valid). Manipulating the QPalette of the `QListWidget`
does not work well enough in the presence of Styles or StyleSheets to be
a viable means of hiding the text underneath either.

Also in the `dlgConnection` class:
* refactor out some highly repetitive code that sets up the predefined
Mud icons.
* remove warnings and adopts a change introduced into Qt 5.15 for
the `QLabel::pixmap()` which now requires a `Qt::ReturnByValue` argument
to prevent an obsolete code warning and to force a QPixmap to be returned
by value and not reference...
* remove a warning about:
  `QListWidget::setItemSelected(QListWidgetItem*,bool)` being deprecated
  and switches to the recommended:
  `QListWidgetItem::setSelected(bool)`

Take the display of the name from the custom icon - as we now show it
visibly underneath and if the user assigns their own icon the text is not
shown anyway.

Change the custom icon generator so that there is NOT a dependency on the
position in the list for the colours used. This means that the same name
will always have the same colours - which will help the user to pick out
a particular profile as it will always look the same no matter if other
profiles are added or removed.

Change the colour gradient to be a radial one rather than a linear one,
this makes it more *interesting* I think.

Refactor out the colour icon generator to a single method that is called
from more than one place.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>