Fixup: insert missing test in TTextCodec methods
The recent re-enabling of warnings revealed a failure to handle a check
I was making in some methods that override the `[pure virtual protected]`
method in:
`(QString) QTextCodec::convertToUnicode(const char *chars, int len,
                                  QTextCodec::ConverterState *state) const`
in some codecs I implimented in the `TTextCodec` class.

As it happens I do not think the method is currently used by Mudlet, but it
makes sense to fix it in case anyone uses those codecs for their own
purposes. (I recall searching for way to do these codecs and none of the
examples I found online were as complete as the ones I have constructed
here!)

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>