Revise: accept peer request - put text back on icons & shrink it under them
This should almost restore the original arrangement (though I am not going
to bother to try setting the text explicitly to white as that just doesn't
work for multiple reasons).

This still means that there are some (coding) enhancements in place - as
well as the slightly funky gradients on custom icons.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>