Enhance: make TConsole scrolling smoother
Make it scroll by only 1 line of text per mouse wheel click instead of
three lines. To compensate for the slower nature of the scrolling now, make
the scrolling faster (to a half the number of lines in the upper
`TTextEdit` pane per step instead of one line) when the control key is
pressed during the wheel event; this thus mimics the same sort of speed
up in the map zoom and scrolling through the selected room number/name
listing in the 2D mapper.

This should close #4437 .

It also fixes a potential bug in that the `TTextEdit::updateScreenView()`
looked to be sending the "screenheight" for both the upper and lower panes
to the same variable in the `Host` class - when they are not going to be
the same value.

It also removes the use of the obsolete `(int) QWheelEvent::delta()`
in favour of the recommended `(QPoint) QWheelEvent::angleDelta()` in the
`qlWidget` class.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>