Cleanup: replace obsolete in Qt 5.15 QComboBox::AdjustToMinimumContents
This `enum` value has been deprecated and `QComboBox::AdjustToContents`
seems to be the best alternative. There was a second case where
`QComboBox::AdjustToMinimumContentsLengthWithIcon` was used (which is still
valid) but since there is no icon used in that case it seemed sensible to
change it to the other usable value as well.

This seems to reduce the warnings by about 5.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>