Revise: select profiles by typing the name, keep profile icons (#4440)
Restore the text display of the profile name.

This is so that the keyboard selection of profiles beginning with a
particular letter works again. This will close issue #4283 .

This also sets it to the minimum size trying to hide it by setting the
font size to 1 - however that still show something (and a font size
of zero is NOT valid). Manipulating the QPalette of the QListWidget
does not work well enough in the presence of Styles or StyleSheets to be
a viable means of hiding the text either.

Also in the `dlgConnection` class:
* refactor out some highly repetitive code that sets up the predefined
Mud icons.
* remove warnings and adopts a change introduced into Qt 5.15 for
the QLabel::pixmap() which now requires a Qt::ReturnByValue argument
to prevent an obsolete code warning and to force a QPixmap to be returned
by value and not reference...
* remove a warning about:
  QListWidget::setItemSelected(QListWidgetItem*,bool) being deprecated
  and switches to the recommended:
  QListWidgetItem::setSelected(bool)

Edit: the last item above may have become redundant as further commits
amalgamated into this PR may have taken out these lines anyhow!

Change the custom icon generator so that there is NOT a dependency on the
position in the list for the colours used. This means that the same name
will always have the same colours - which will help the user to pick out
a particular profile as it will always look the same no matter if other
profiles are added or removed.

Remove  the colour gradient in the custom icon as it is not thought to be
useful nowadays.

Refactor out the colour icon generator to a single method that is called
from more than one place.

BugFix: restore odd original way of accessing some QPixmaps

This was causing failures in some CI platforms using Qt 5.14. or
older.

Revise: really frob the custom icon generator

This is to try and produce as varied but still readable icons when they are
generated entirely on the basis of the MUD name - in a reproducible
(though Qt do NOT guarantee repeatable qHash() functionality across Qt
versions - so they may not stay *that* constant) manner.

Refrob: use a table of fixed colours

Using only 15 hues but with 3 different saturation level plus 5 grey-scale
one should give enough different icons but not generate them too close to
each other for those with normal colour vision...

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>

Co-authored-by: Kebap <kebap_spam@gmx.net>