Cleanup: replace obsolete in Qt 5.15 QComboBox::activated(const QString&)
This (overloaded) signal has been obsoleted and we must use the other
(overload) that returns the index of the activated item rather than the
text that we were using. For the one in the 2D mapper this is slightly
more fiddly as we actually need to pass the text to another class whereas
the editor case is simpler as it is used within the same class so we
can use the alternate form almost as easily.

This should reduce the warnings by one or two for Qt 5.15

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>