Cleanup: adopt "Rule of Three" requirements for TChar class
Briefly: since we have a user-defined copy constructor we need to also
have an assignment operator and a destructor - though they can be default
ones.

I discovered this when getting a warning when using the Clang 10 compiler
on FreeBSD and found out more about it on Stack Exchange at:
https://stackoverflow.com/q/51863588/4805858

Also:
* Because the action of the previous default constructor and the one that
takes a `Host*` is the same if the latter is give a `nullptr` argument the
former is redundant and can be removed if we default the `Host*` argument
for the latter and do NOT mark it as `explicit`.
* remove the std::move(...) wrapper at end of `TBuffer::log(int, int) as it
is causing Clang 10 to warn:
"Moving a temporary object prevents copy elision [-Wpessimizing-move]"

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>