Cleanup: remove unused class members in LuaInterface and XMLimport
The `Host*` is never actually used inside the `LuaInterface` class, only in
a non-default constructor. So it does not need to be retained as a class
member.

As the default is never used either we can also mark the above constructor
as `explicit`...

Also `mMaxAreaId` is declared in the `XMLimport` class but although a use
had been considered it is not currently used for anything.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>