Cleanup: remove undesirable unsigned to signed integer comparision
This originally was just going to convert a use of `size_t` to hold the
result of a Qt class `::size()` when they use `int` - so as to avoid a
warning about comparing signed and unsigned integers. However I spotted
some other "improvements" that could be made in the same class by using
`::at(int)` instead of `::operator[int]`; by marking unchangable things
as `const` and using pre-increment rather than the slower post-increment
operator.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>