Cleanup: remove unused retention of method return value
Since we do not use it in this case there is no point in storing it, as
this is prompting an unused variable and there is not really a conceivable
use for this particular case. The method is called in two other places and
the result used in only one of them as well!

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>