Cleanup: replace deprecated in Qt 5.15 date format methodology
Qt pointedly makes it clear that many of the formats supported by
`QDate::toString(...)` will be removed in Qt 6 and that the
`Qt::DateFormat` `Qt::LocalDate` should be replaced by the `QString` that
`QLocale::dateFormat(QLocale::ShortFormat)` returns for the system local.

As I have arranged for the `mudlet` class to have a `getUserLocale()`
method that can return a `QLocale` to match that which the user has set to
be used to customise it to their Language preferences it makes sense to use
that to determine the actual (short) date format string to use.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>