Cleanup: replace obsolete in 5.11 use of QDesktopWidget
The whole class was declared obsolete before Qt 5.11. It seems that one is
supposed to consider instead the data retrievable from `QList<QScreen*>`.
Annoyingly it is made clear that `QScreen::availableGeometry()` does not
produce reliable results on X11 due to a limitation on the specification.

Also:
* As of Qt 5.15 it is possible to direct the screen (monitor) to which a
Splash-screen is applied, I noticed that this was not always to the same
monitor to which the main Mudlet window was sent.  I have arranged for them
now to both be applied to the primary screen, which definitely works for
main application window but I am not so sure the splash screen follows the
instruction...

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>