Cleanup: replace obsoleted in Qt 5.15 QSet:toList()
This seems to have been missed in a previous cleanup, perhaps because it is
being used on the output of a `QHash` with values that are of the form
of `QSet<int>`. The Qt documentation does say: "Note: Since Qt 5.14, range
constructors are available for Qt's generic container classes and should be
used in place of this method." so it may be that a further change will be
needed but since I have already had one or two run ins with those range
constructors I do not feel the urge to dive in on that right now.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>