Revise: suggestion to put default mouse scroll to 3 lines
That is 3 lines per default (15 degree) click of the (vertical) scroll
wheel. To get the single line per click it can now be obtained by using the
shift key, and that overrides the other override of half a screen height of
lines that the control key gives.

This might close the resurrected #4437.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>