BugFix: correct QRegularExpression warnings on profile save
This was causing messages of the form:
"'QRegularExpressionPrivate::doMatch(): called on an invalid
QRegularExpression object'" whenever a color trigger was saved. It was due
to an error in the regular expression defined in:
`QStringList XMLexport::remapAnsiToColorNumber(const QStringList&,
                                                        const QList<int>&)`

Also remove unneeded `#include <QRegularExpression>`  in classes that do
not need it.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>