Cleanup: various things for Keys item type (#4466)
Convert internals to store and use the Qt `enum`s dedicated to keyboard
handling (`Qt::Keys` and `Qt::KeyboardModifiers`). Means we can justify
using the enum values directly.

Report the keycode (in hex) for a key that isn't in our lookup table of
names.

Removes some C-style casts (to `int` type).

Renames a `QKeyEvent` argument in some places from `key` to `ke` as the
former might be confusingly misread as a key code not an event.

Improve the handling of <Ctrl>+number key to switch profiles (it adds
handling <Ctrl>+0 for a tenth profile. By a careful bit of refactoring it
has been possible to eliminate some "fall-throughs"in a `switch` so that
up to 8 redundant calls to `TCommandLine::handleCtrlTabChange(...)` in a
row when <Ctrl>+1 is used - and to remove the warnings about those
fallthroughs.

Remove a whole load of else-after-return code anti-patterns.

Simplify a QMap value lookup to use the QMap::value() directly without the
redundant QMap::find() != QMap::end() first.

I estimate this should reduce the number of warnings by about 7.

I recognise that this will interfere with my stalled (um, closed) pull
request #2411 which I will have to revise to accommodate this one!

Revised to try and shorten a lengthy engineer's comment to translators
which was suggested during peer review (it was a tad too long)...

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>