BugFix: repair incorrect use of QFont(string) after QFont::toString()
Creating a font via a `QFont(string)` where `string` is the output from
a previous `QFont::toString()` is incorrect as each time it is done it
adds on the last nine of the ten comma-separated elements on in a manner
that the string will grow interminably. This is done for the Map font when
using the fall-back (pre-format 19) method and as an example I had one map
file that had over 15K comma separated elements.

The correct way to use such a string is with the
`QFont::fromString(string)` method - however that will reject a string that
has other than one element or ten comma separated ones. This PR will fix
up any cases it encounters.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>