BugFix: correct numerical parameter corruption in event handler (#4524)
Large numbers were losing *precision* when passed as event parameters.

It turns out that the default conversion to a `QString` done in
`QString::number(double)` is not passing as many significant digits as some
users want when it is used in the Lua `raiseEvent(...)` function. That
method actually has a couple of defaulted parameters:
`QString::number(double value, char format = 'g', int precision = 6)`
and we want more than 6 digits for our uses. The maximum that seems to work
is 17 so use that.

Also, the process of identifying what the arguments are that are passed to
`raiseEvent(...)` were using the `lua_isxxx(...)` functions - however those
will flag something that can be coerced into the tested for type. In this
situation it is probably better to actually see what the type is thought to
be, so instead use the `lua_type(...)` function and examine the return
value.

As largely the same code is present in the `raiseGlobalEvent(...)` function
this PR does the same things there as well.

This will close #4502.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>