W.I.P. start on reader
Renamed `(QMap<int, int>) TMap::envColors` to `TMap::mEnvColors`
and `(QMap<int, QColor>) Tmap::customEnvColors` to `TMap::mCustomEnvColors`
so that it is clearer that they are members of the `TMap` class.

Add alpha component to end of list of (now four) 0 to 255 integer values
returned by `getCustomEnvColorTable()` - as the corresponding setter does
allow one to be provided.

Extract some of the `QStringLiterals` in `./src/TMap.cpp` to predefined
constants so that only one of each string is created/needed in that file.

Fixed handling of `TMap::mCustomEnvColors` as it was not being written
correctly. Version incremented to `0.002` - and only that form will be
read currently. The reader is incomplete at present - only a test read of
the direct TMap elements are loaded to temporary variables for testing.
The area (and their contained rooms) are still to have parser created.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>