Cleanup: apply previously version dependent fix to all (#4580)
It seems that the change I proposed in #4461 does not apply at the correct
Qt version breakpoint. I had set the newer code to be used from Qt 5.15.
However it is still throwing up GH Actions "check" annotations which must
therefore be testing with an older Qt version. Looking more deeply into
the Qt code it seems that the particular method was deprecated in 5.8 and
since we only now support Qt versions going back to 5.11 there is no need
to keep the old code.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>