Revise: refactor TTextEdit text painting code to improve clipping
The existing code draws the foreground and the background of each glyph in
a line of text at the same time - with the negative effect that the
background painting will overpaint the foreground of preceding glyphs.
This is unwanted; this PR revises things so that the entire backgrounds for
all the glyphs in a line are painted before the foregrounds are painted
on top.

Remove:
`(void) drawBackground(QPainter&, const QRect&, const QColor&) const`
and insert its couple of lines back into the two places it was used.

Split the `(int) drawGrapheme(...)` code into a version that does the back
ground and the foreground in two separate passes.

Move the `(int) drawGrapheme(...)` into two separate methods and refactor
the code in each so that the first used one:
`(int) drawGraphemeBackground(...)`
does the processing and stores the intermediate bits of data in a form that
the second one:
`void drawGraphemeForeground(...)`
can use.

This should help with issue #4609.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>