BugFix: improve handling of TConsole::mSystemMessageBgColor
This `QColor` is initilised to the default value of `TConsole::mBgColor`
which, since #4570, is now the transparent color. Unfortunely this is bad
because when text printed on the main console with:
`(void) TConsole::printSystemMessage(const QString)`
the red foreground and the transparent background get swapped over and thus
is rendered as a solid block of red.

This PR attempts to solve this by ensuring that whne the
`TConsole::mBgColor` member is set to a different color that the same
colour is also stored for use as `TConsole::mSystemMessageBgColor`.

The issue of how things are colored does seem to be made more complicated
by the use of `QPalette` for some things for which it does not seem to be
worthwhile; therefore in this PR I also remove some uses of the Qt palette
system in places where it does not seem to be doing anything useful.

I have also reviewed how things might be affected by setting an image as
a background image for the main console but it seems that the changes here
do not prevent the display of an image where there is NO text shown on
screen.

This might close #4635.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>