Remove TLS Default for StickMUD (#4672)
Although the TLS connection for StickMUD works, there is more work to do on the StickMUD side with keeping the connection from timing out after a number of minutes. Until this is stable, we need to tell players to change the port and uncheck secure, which is not the best experience.