BugFix: correct mis-use of nullptr to denote no room symbol colour
It seems like PR #4573 introduced a bogus use of pointer access to a
`QColor` whereas the correct methodology is `QColor::isValid()` - a default
constructed `QColor` is itself equivalent to an RGBA specified colour of
`QColor(0,0,0,0)` and that will fail in that test. This PR fixes things so
that (instead of a `nullptr` test) is used.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>