BugFix: initialise mDlgRoomSymbol so that settng room symbols works again
A test for this pointer being a `nullptr` is what enables a new dialogue to
be made in the 2D mapper - by not initialising it in #4573 a random
non-null value is found so the dialogue and thus the functionality can be
virtually impossible to use.

As it is a pointer I have taken the liberty to introduce a `p` into the
initial part of the name - after the `m` to indicate it is a class member.

As the defect is a regression that causes loss of functionality I am rating
this as a **high** priority bug(fix) even though it does not induce any
fatal termination of the application.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>