BugFix: correct merge error from #4604
Due to some differences in local variable names I did not spot that I was
introducing two copies of the same functional code into the method that
scanned the existing map file in other inactive profiles for when a new
shared map was copied over to them. As such the code (in
`(bool) TMap::retrieveMapFileStats(...)` would not have been used very
much in the last couple of weeks when it was in the development branch code
and even then it may not have prevented the current usage of that method
from having functioned correctly. Nevertheless it is defective as it is
and this PR removes one copy of the duplicated code.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>