Fixup: ensure code used in this PR matches the development branch version
In creating #4693 I chose the opposite duplicate chunk of code that I was
using in this PR - since there was slight differences it would have
impacted when this PR is merged.

This commit swaps things so that it uses the same code as will be in
the development branch in the near future.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>