Revise: add distinctive prefix to all the new static const values
During a CI build of a previous commit it failed on one platform because
one of the `static const` I had used `SIZE` was clashing with something
else. Also I was noticing that `ID` was lighting up in my Linux `qtcreator`
as if it was a member of one of the map classes - yet it was not one that
I could see defined anywhere. To avoid mixups I have added a `sc` for
`static const` prefix - which does fit with some of the other prefixes we
are using.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>