BugFix: correct missing detail (map label position) from binary map files
This will cause serious data loss (because it throws off the whole
structure of the file) when reading existing files {because the label
position element is not read from the file so everything after the first
map label will not be at the expected offset} and writing new ones as they
would not be storing the position detail rendering all map labels at the
origin when read by the new code.

This defect was introduced in #4604 - ten days ago.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>