BugFix: correctly evaluate room m(in|ax)(x|y) values in all cases
We previously set them all to be 0.0 when there is no custom lines but
technically, since they are supposed to be extremes of the room and it's
custom exit lines they should be initialised to the room's own coordinates.

Not doing this meant that in some cases they were being used but they
did not hold sensible values.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>