Revise: include a timeout on concatenating keypresses
This is to provide the same functionality that the QListWidget does for
keypresses, if the keys are typed fast enough they are concatenated to form
the text to search for. However if the user types slower than this then
the individual keys are only considered on their own. This reflects what
both Vadim and I experienced - in that he was typing fast enough to spell
out a particular profile but I wasn't and so the profile that I was having
selected was changing on each keypress.

The time between keypresses allowed so that they can be concatenated I have
set to be 1 second. This is significantly slower that the Qt code but it
is long enough to permit the temporary disablement of non-matching profiles
to be a useful visual feedback - and besides, it makes it easier to specify
one of a number of profiles with a common initial text - the longer the
timeout the more keys the user can string together!

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>