Enhance: show profile indication in central debug console output
This PR adds a single character inside `[`...`]` to the beginning of
(nearly) all texts displayed on the Central Debug Console which is made
unique for each session of a profile run in a Mudlet application run.
This marking (except for messages not from a particular profile) is not
displayed when there is only one profile - and the start and end (and
a change of name) of each profile is noted and a revised listing of the
symbol used for each active profile is shown each time it changes.

To do this the previous dummy `>> 0` used at the end of a `TDebug()`
message creation has been changed to instead take a pointer to a profile's
`host` instance - from which the profile name is extracted. To allow for
multi-coloured messages to be concatenated, inserting a
`(QChar) TDebug::csmContinue` element at the start of a new `TDebug` text
will cause it to be appended onto the end of a previous `TDebug` text
without acquiring the profile indication marking.

Messages that are generated with a `nullptr` as the Host pointer are
expected to be messages from the Mudlet Application rather than a
particular profile and they will always be marked with a `[*]`.

When multi-playing closing a profile and reopening it again will cause a
new identifier to the used for the subsequent run - so it can be seen
which run any messages are from.

The current design uses identifiers from the sequence:
'1' to '9', then 'A' to 'Z', then 'a' to 'z' - after that set of 61 marks
are exhausted then all subsqeuent profile runs in that Mudlet application
run are shown with a '?'. it would be possible to extend the sequence but
as the Central Debug console is intended to help with debugging it is not
expected for the finite number of identifiers to be a serious limitation!

This will close #804 (originally a LaunchPad issue:
https://bugs.launchpad.net/mudlet/+bug/1658271 )!

Also:
* simplify some Lua handling of boolean returns and conditions based upon
an integer value.
* in `TTrigger::match(...)` use `QStringList::at(...)` in some places where
it is faster/safer than `QStringList::operator[]()`.
* Remove a rogue line-feed prior to a:
"LUA: code compiled without errors. OK"
message which was making it more spammy than it needed to be.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>